refactor(desktop): extract markdown preview to shared util
Move the unified+remarkParse+mdastToString preview helper out of the
v2 review tab into renderer/utils/markdownPreview so future v2 spots
that need a one-line plain-text excerpt of markdown can reuse it.

v1 ReviewPanel keeps its existing regex extractor — sunset path,
not worth touching.
