+++ apps/desktop/src/renderer/utils/markdownPreview/index.ts
@@ -1,0 +1,1 @@
+export { getMarkdownPreviewText } from "./markdownPreview";
